refactor(agent): extract pure compaction helpers (#3250)
Behavior-preserving extraction of the pure compaction cluster from
agent.store.ts (#3250) into a co-located module src/lib/agent/compaction.ts.

Moved (no `this.`, no closure over store state — pure functions/class/types):
- PredictiveCompactMutex (class)
- prunableAgentMessage, applyPrunedAgentMessages, buildAgentCompactionPrepend
- defaultContextWindowFor (+ its private CLAUDE_1M_TIER_CAPABLE_MODELS set)
- CompactionOutcome, CompactAgentResult (types)

AgentMessage / PrunableMessage are type-only imports (erased at runtime, no
cycle). The predictiveCompactMutex singleton stays in the store (instantiates
the imported class). Every call site is unchanged; the store re-exports
PredictiveCompactMutex and CompactionOutcome so its public surface is
identical. The scalar constants (BUDGET_*, SUMMARY_*, AGGRESSIVE_RETRY_*,
PREDICTIVE_*) and the state-closing helpers (consumeCompactionPrepend,
waitForStandbySeed, drainStandbyQueueIfPending) stay in the store.
agent.store.ts drops from 9281 to 9101 lines.

Six tests pin the moved declarations by SOURCE TEXT
(readFileSync/readSource of agent.store.ts). Those specific assertions are
repointed to the compaction.ts source where the code now lives — same fact,
same type/body asserted, zero weakening. Behavioral call-site assertions on
compaction methods that remain in the store stay pointed at agent.store.ts.

No behavior, IPC-contract, or persistence-schema change.

Validation: full unit suite green (2714 passed); Biome clean; all changed
files type-clean under tsc.

Taariq Lewis, SerenAI, Paloma, and Volume at https://serendb.com
Email: hello@serendb.com
